requirements.txt
# Core RAG pipeline
langchain==0.1.14
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.40.0

# LLM and embeddings
groqcloud==0.4.0           # Official Groq Python SDK
huggingface-hub==0.21.4
accelerate==0.27.2

# Document loading
pypdf==3.13.0
python-docx==0.8.11

# Web frontend
streamlit==1.34.0

# Utility
tqdm==4.66.4
python-dotenv==1.0.1

# Optional MLflow (if you want tracking later)
mlflow==2.12.1
